[Admin][ProductVariant] Rename MovePositionsController to SavePositionsController
