+++ src/Sylius/Bundle/AdminBundle/Resources/assets/controllers/SavePositionsController.js
