+++ src/Sylius/Bundle/AdminBundle/Resources/views/ProductVariant/Grid/Action/updatePositions.html.twig
@@ -1,13 +1,13 @@
 {% from '@SyliusAdmin/Shared/Helper/icon.html.twig' import icon %}
 
 <button class="btn primary"
-    {{ stimulus_controller('move-positions', {
+    {{ stimulus_controller('save-positions', {
         url: path('sylius_admin_ajax_product_variants_update_position'),
         csrfToken: csrf_token('update-product-variant-position'),
         inputSelector: '.sylius-product-variant-position',
         dataKey: 'productVariants',
     }) }}
-    data-action="move-positions#submit"
+    data-action="save-positions#submit"
     {{ sylius_test_html_attribute('save-configuration-button') }}
 >
         {{ icon({ icon: action.icon ?? 'check', class: 'icon icon-tabler dropdown-item-icon' }) }}
