+++ src/Sylius/Bundle/AdminBundle/Resources/views/ProductVariant/Grid/Field/position.html.twig
@@ -1,1 +1,1 @@
-<input type='text' value="{{ data.position }}" data-id="{{ data.id }}" class="sylius-product-variant-position form-control text-center"/>
+<input type='text' value="{{ data.position }}" data-id="{{ data.id }}" class="sylius-product-variant-position form-control text-center" />
